requirements.txt
flask==3.1.3
gunicorn==23.0.0
pytest==8.3.0
pytest-cov==5.0.0
flake8==7.1.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
